+++ src/visualize.py
@@ -38,7 +38,7 @@ class Visualizer:
         plt.show()
 
     @classmethod
-    def display_rois(cls, image, feature_maps, rois):   # TODO: There appears to be a shift between the image and feature map plots
+    def display_rois(cls, image, feature_maps, rois):
         fig, axes = plt.subplots(1, 2, figsize=(10, 5))
 
         image = cls._convert_to_original_image(image)
@@ -49,12 +49,18 @@ class Visualizer:
         scale_x = image.shape[0] / feature_maps.shape[1]
         scale_y = image.shape[1] / feature_maps.shape[2]
 
+        fm_offset = -0.5 # Offset to align the ROIs with the feature map due to a shift that plt produces
+
         for i, roi in enumerate(rois):
             x1, y1, x2, y2 = roi
             color = DEFAULT_COLORS[i % len(DEFAULT_COLORS)]
 
             # Scale the coordinates to the image shape
-            ix1, iy1, ix2, iy2 = x1 * scale_x, y1 * scale_y, x2 * scale_x, y2 * scale_y
+            ix1, ix2 = x1 * scale_x, x2 * scale_x
+            iy1, iy2 = y1 * scale_y, y2 * scale_y
+
+            # Apply the offset
+            x1, y1, x2, y2 = x1 + fm_offset, y1 + fm_offset, x2 + fm_offset, y2 + fm_offset
 
             axes[0].add_patch(plt.Rectangle((ix1, iy1), ix2 - ix1, iy2 - iy1, fill=False, edgecolor=color, lw=1))
             axes[1].add_patch(plt.Rectangle((x1, y1), x2 - x1, y2 - y1, fill=False, edgecolor=color, lw=1))